Move apy and apr to src/math
